rename filter dictionary
